stacks/dagster/requirements.txt
# --- Dagster core ---
dagster
dagster-webserver
dagster-postgres
dagster-dbt

# --- Database ---
psycopg2-binary

# --- HTTP / data ---
requests
resend
pandas==2.1.4
numpy==1.26.4
pyarrow==16.1.0

# --- Web scraping (pfn depth chart) ---
beautifulsoup4
lxml

# --- Mail intelligence ---
anthropic
pdfplumber
PyMuPDF
pytesseract
Pillow
boto3

# --- dbt (Fusion binary installed in Dockerfile; adapter package still required for DB connections) ---
dbt-postgres
